requirements.txt
matplotlib
scipy
pillow
opencv-python
opencv-contrib-python
open3d
trimesh
scikit-learn
scikit_video
pybullet
joblib
tqdm
termcolor
pyyaml
GitPython
tensorboard
deprecation
numpy==1.23
ipdb
chumpy
manotorch
rtree